feat(dcp-evaluator-start): change how script behaves for node evaluator
To interface with localExec nicely, the script has been changed to
output a JSON array of set up files if the evaluator type is set to
node. It makes it easier for programs to parse the output of the script.

Also fixed a bug where process.stdout.clearLine and
process.stdout.cursorTo were not defined by importing the functions
from the built-in `readline` module.