ci: improve caching, conditions for running, & inheritance
These changes are primarily motivated by making the addition of new jobs
easier.

Use of the `cache` keyword is added to make adding jobs that depend on
`node_modules` easier, along with reducing the amount of work `npm`
does during a clean install.

Pipelines run less often now since the results of branch pipelines
aren't useful to the team's workflow.

The inheritance from the `default` job was updated with the
`inherit:default` keywords to try and make it clearer what the relevant
parts of the `default` job are to certain CI jobs.

Refs: https://docs.gitlab.com/ee/ci/variables/predefined_variables.html
Refs: https://docs.gitlab.com/ee/ci/yaml/#inheritdefault